Segregated DNA/GA from specs.py
